requirements.txt
et-xmlfile==1.1.0
Faker==18.4.0
openpyxl==3.1.2
psycopg2==2.9.6
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
